Main activity has been created.
It includes following ViewPager and Toolbar.

Tabs are created with 3 Fragments.

These fragments used with adapter which is PagerAdapter for ViewPager.
