+++ app/src/main/AndroidManifest.xml
@@ -10,8 +10,11 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
+            android:name=".ProfileActivity"
+            android:label="@string/app_name" />
+        <activity
             android:name=".MainActivity"
-            android:label="@string/app_name" >
+            android:label="@string/title_activity_main">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
