+++ app/src/main/res/values/strings.xml
@@ -23,5 +23,14 @@
         <item>Select from gallery</item>
     </string-array>
 
+    <string name="start">START</string>
+    <string name="history">HISTORY</string>
+    <string name="settings">SETTINGS</string>
+
+    <string name="title_activity_main">MainActivity</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+
 </resources>
